+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `aad-pod-identity` - updated chart to 4.1.7 [@sossickd](url)
 - `cert-manager` - added toleration and node selector for startupapicheck [@sossickd](url)
+- `kubectl provider` - enabled server-side-apply aad-pod-identity [@sossickd](url)
 
 <br>